fixed actor critic

+++ custom_rl_jax/vec_policy_gradient_cs/metrics/finished_reward_recorder.py
@@ -31,7 +31,7 @@ class FinishedRewardRecorder:
         ongoing_episode_rewards = state["ongoing_episode_rewards"] + step_rewards
 
         finished_episode_rewards = jnp.where(done, ongoing_episode_rewards, state["finished_episode_rewards"])
-        ongoing_episode_rewards = jnp.where(done, jnp.zeros((self.vec_num,)), ongoing_episode_rewards)
+        ongoing_episode_rewards = jnp.where(done, jnp.zeros((self.vec_num,), dtype=jnp.float32), ongoing_episode_rewards)
 
         return {
             "finished_episode_rewards": finished_episode_rewards,
