+++ custom_rl_jax/vec_policy_gradient_cs/metrics/metrics_logger_np.py
@@ -16,12 +16,12 @@ class MetricsLoggerNP:
     def __init__(self, total_steps: int):
         self.total_steps = total_steps
         self.curser = 0
-        self.mean_rewards = np.zeros(total_steps)
-        self.state_value = np.zeros(total_steps)
-        self.td_error = np.zeros(total_steps)
-        self.actor_loss = np.zeros(total_steps)
-        self.critic_loss = np.zeros(total_steps)
-        self.entropy = np.zeros(total_steps)
+        self.mean_rewards = np.zeros(total_steps, dtype=np.float32)
+        self.state_value = np.zeros(total_steps, dtype=np.float32)
+        self.td_error = np.zeros(total_steps, dtype=np.float32)
+        self.actor_loss = np.zeros(total_steps, dtype=np.float32)
+        self.critic_loss = np.zeros(total_steps, dtype=np.float32)
+        self.entropy = np.zeros(total_steps, dtype=np.float32)
 
     def log(self, metrics_frame: MetricsRecorderState):
         frame_length = len(metrics_frame['mean_rewards'])
